+++ TrafficMonitor/NetworkInfoDlg.h
@@ -12,7 +12,7 @@ class CNetworkInfoDlg : public CBaseDialog
     DECLARE_DYNAMIC(CNetworkInfoDlg)
 
 public:
-    CNetworkInfoDlg(vector<NetWorkConection>& adapters, MIB_IFROW* pIfRow, int connection_selected, CWnd* pParent = NULL);   // 标准构造函数
+    CNetworkInfoDlg(vector<NetWorkConection>& adapters, MIB_IFROW* pIfRow, size_t connection_selected, CWnd* pParent = NULL);   // 标准构造函数
     virtual ~CNetworkInfoDlg();
 
     // 对话框数据
@@ -26,8 +26,8 @@ protected:
 
     vector<NetWorkConection>& m_connections;
     MIB_IFROW* m_pIfRow;
-    int m_connection_selected;		//当前对话框显示的连接
-    int m_current_connection;		//初始选择的连接
+    size_t m_connection_selected;		//当前对话框显示的连接
+    size_t m_current_connection;		//初始选择的连接
 
     CListCtrlEx m_info_list;
     CMenu m_menu;
